Allow passing Ints as error codes to "exit".
Actually all exit codes should be of type Int instead of Int32. But changing it in the rest of the framework could break backwards compatibility.

+++ Sources/SwiftShell/Context.swift
@@ -78,7 +78,7 @@ private func createTempdirectory () -> String {
 		try Files.createDirectoryAtPath(tempdirectory.path!, withIntermediateDirectories: true, attributes: nil)
 		return tempdirectory.path! + "/"
 	} catch let error as NSError {
-		exit(errormessage: "Could not create new temporary directory '\(tempdirectory)':\n\(error.localizedDescription)", errorcode: Int32(error.code))
+		exit(errormessage: "Could not create new temporary directory '\(tempdirectory)':\n\(error.localizedDescription)", errorcode: error.code)
 	} catch {
 		exit(errormessage: "Unexpected error: \(error)")
 	}
